@@ -4,21 +4,21 @@
 from __future__ import annotations
 
 import bdb
+from collections.abc import Callable
+from collections.abc import Generator
+from collections.abc import Iterable
+from collections.abc import Sequence
 from contextlib import contextmanager
 import functools
 import inspect
 import os
 from pathlib import Path
 import platform
+import re
 import sys
 import traceback
 import types
 from typing import Any
-from typing import Callable
-from typing import Generator
-from typing import Iterable
-from typing import Pattern
-from typing import Sequence
 from typing import TYPE_CHECKING
 import warnings
 
@@ -353,7 +353,7 @@ class DoctestItem(Item):
                 # add line numbers to the left of the error message
                 assert test.lineno is not None
                 lines = [
-                    "%03d %s" % (i + test.lineno + 1, x) for (i, x) in enumerate(lines)
+                    f"{i + test.lineno + 1:03d} {x}" for (i, x) in enumerate(lines)
                 ]
                 # trim docstring error lines to 10
                 lines = lines[max(example.lineno - 9, 0) : example.lineno + 1]
@@ -593,7 +593,6 @@ class DoctestModule(Module):
 
 def _init_checker_class() -> type[doctest.OutputChecker]:
     import doctest
-    import re
 
     class LiteralsOutputChecker(doctest.OutputChecker):
         # Based on doctest_nose_plugin.py from the nltk project
@@ -636,7 +635,7 @@ def _init_checker_class() -> type[doctest.OutputChecker]:
             if not allow_unicode and not allow_bytes and not allow_number:
                 return False
 
-            def remove_prefixes(regex: Pattern[str], txt: str) -> str:
+            def remove_prefixes(regex: re.Pattern[str], txt: str) -> str:
                 return re.sub(regex, r"\1\2", txt)
 
             if allow_unicode:
