@@ -855,7 +855,7 @@ class TestLocalPath(CommonFSTests):
         assert b.fnmatch(pattern)
 
     def test_sysfind(self):
-        name = sys.platform == "win32" and "cmd" or "test"
+        name = (sys.platform == "win32" and "cmd") or "test"
         x = local.sysfind(name)
         assert x.check(file=1)
         assert local.sysfind("jaksdkasldqwe") is None
@@ -948,7 +948,7 @@ class TestExecution:
                 prefix="base.", rootdir=tmpdir, keep=2, lock_timeout=0
             )
             assert numdir.check()
-            assert numdir.basename == "base.%d" % i
+            assert numdir.basename == f"base.{i}"
             if i >= 1:
                 assert numdir.new(ext=str(i - 1)).check()
             if i >= 2:
@@ -993,7 +993,7 @@ class TestExecution:
         for i in range(10):
             numdir = local.make_numbered_dir(prefix="base2.", rootdir=tmpdir, keep=2)
             assert numdir.check()
-            assert numdir.basename == "base2.%d" % i
+            assert numdir.basename == f"base2.{i}"
             for j in range(i):
                 assert numdir.new(ext=str(j)).check()
 
@@ -1250,7 +1250,7 @@ class TestWINLocalPath:
     def test_chmod_simple_int(self, path1):
         mode = path1.stat().mode
         # Ensure that we actually change the mode to something different.
-        path1.chmod(mode == 0 and 1 or 0)
+        path1.chmod((mode == 0 and 1) or 0)
         try:
             print(path1.stat().mode)
             print(mode)
