@@ -5,10 +5,10 @@ and by builtin plugins."""
 
 from __future__ import annotations
 
+from collections.abc import Mapping
+from collections.abc import Sequence
 from pathlib import Path
 from typing import Any
-from typing import Mapping
-from typing import Sequence
 from typing import TYPE_CHECKING
 
 from pluggy import HookspecMarker
