@@ -10,14 +10,13 @@ https://github.com/jenkinsci/xunit-plugin/blob/master/src/main/resources/org/jen
 
 from __future__ import annotations
 
+from collections.abc import Callable
 from datetime import datetime
 from datetime import timezone
 import functools
 import os
 import platform
 import re
-from typing import Callable
-from typing import Match
 import xml.etree.ElementTree as ET
 
 from _pytest import nodes
@@ -48,7 +47,7 @@ def bin_xml_escape(arg: object) -> str:
     The idea is to escape visually for the user rather than for XML itself.
     """
 
-    def repl(matchobj: Match[str]) -> str:
+    def repl(matchobj: re.Match[str]) -> str:
         i = ord(matchobj.group())
         if i <= 0xFF:
             return f"#x{i:02X}"
