@@ -3,10 +3,10 @@
 from __future__ import annotations
 
 import collections
+from collections.abc import Collection
+from collections.abc import Iterable
+from collections.abc import Set as AbstractSet
 import dataclasses
-from typing import AbstractSet
-from typing import Collection
-from typing import Iterable
 from typing import Optional
 from typing import TYPE_CHECKING
 
