@@ -2,17 +2,17 @@
 from __future__ import annotations
 
 import abc
+from collections.abc import Callable
+from collections.abc import Iterable
+from collections.abc import Iterator
+from collections.abc import MutableMapping
 from functools import cached_property
 from inspect import signature
 import os
 import pathlib
 from pathlib import Path
 from typing import Any
-from typing import Callable
 from typing import cast
-from typing import Iterable
-from typing import Iterator
-from typing import MutableMapping
 from typing import NoReturn
 from typing import overload
 from typing import TYPE_CHECKING
@@ -143,14 +143,14 @@ class Node(abc.ABC, metaclass=NodeMeta):
     # Use __slots__ to make attribute access faster.
     # Note that __dict__ is still available.
     __slots__ = (
-        "name",
-        "parent",
-        "config",
-        "session",
-        "path",
+        "__dict__",
         "_nodeid",
         "_store",
-        "__dict__",
+        "config",
+        "name",
+        "parent",
+        "path",
+        "session",
     )
 
     def __init__(
