@@ -77,6 +77,7 @@ from _pytest.warning_types import PytestCollectionWarning
 from _pytest.warning_types import PytestConfigWarning
 from _pytest.warning_types import PytestDeprecationWarning
 from _pytest.warning_types import PytestExperimentalApiWarning
+from _pytest.warning_types import PytestFDWarning
 from _pytest.warning_types import PytestRemovedIn9Warning
 from _pytest.warning_types import PytestUnhandledThreadExceptionWarning
 from _pytest.warning_types import PytestUnknownMarkWarning
@@ -88,41 +89,27 @@ set_trace = __pytestPDB.set_trace
 
 
 __all__ = [
-    "__version__",
-    "approx",
     "Cache",
     "CallInfo",
     "CaptureFixture",
     "Class",
-    "cmdline",
-    "Collector",
     "CollectReport",
+    "Collector",
     "Config",
-    "console_main",
-    "deprecated_call",
     "Dir",
     "Directory",
     "DoctestItem",
-    "exit",
     "ExceptionInfo",
     "ExitCode",
-    "fail",
     "File",
-    "fixture",
     "FixtureDef",
     "FixtureLookupError",
     "FixtureRequest",
-    "freeze_includes",
     "Function",
-    "hookimpl",
     "HookRecorder",
-    "hookspec",
-    "importorskip",
     "Item",
     "LineMatcher",
     "LogCaptureFixture",
-    "main",
-    "mark",
     "Mark",
     "MarkDecorator",
     "MarkGenerator",
@@ -131,7 +118,6 @@ __all__ = [
     "MonkeyPatch",
     "OptionGroup",
     "Package",
-    "param",
     "Parser",
     "PytestAssertRewriteWarning",
     "PytestCacheWarning",
@@ -139,31 +125,47 @@ __all__ = [
     "PytestConfigWarning",
     "PytestDeprecationWarning",
     "PytestExperimentalApiWarning",
-    "PytestRemovedIn9Warning",
-    "Pytester",
+    "PytestFDWarning",
     "PytestPluginManager",
+    "PytestRemovedIn9Warning",
     "PytestUnhandledThreadExceptionWarning",
     "PytestUnknownMarkWarning",
     "PytestUnraisableExceptionWarning",
     "PytestWarning",
-    "raises",
+    "Pytester",
     "RecordedHookCall",
-    "register_assert_rewrite",
     "RunResult",
     "Session",
-    "set_trace",
-    "skip",
     "Stash",
     "StashKey",
-    "version_tuple",
-    "TempdirFactory",
     "TempPathFactory",
+    "TempdirFactory",
     "TerminalReporter",
-    "Testdir",
     "TestReport",
     "TestShortLogReport",
+    "Testdir",
     "UsageError",
     "WarningsRecorder",
+    "__version__",
+    "approx",
+    "cmdline",
+    "console_main",
+    "deprecated_call",
+    "exit",
+    "fail",
+    "fixture",
+    "freeze_includes",
+    "hookimpl",
+    "hookspec",
+    "importorskip",
+    "main",
+    "mark",
+    "param",
+    "raises",
+    "register_assert_rewrite",
+    "set_trace",
+    "skip",
+    "version_tuple",
     "warns",
     "xfail",
     "yield_fixture",
