@@ -1,10 +1,10 @@
 # mypy: allow-untyped-defs
 from __future__ import annotations
 
+from collections.abc import Generator
 import dataclasses
 import re
 import sys
-from typing import Generator
 
 from _pytest.monkeypatch import MonkeyPatch
 from _pytest.pytester import Pytester
@@ -119,8 +119,8 @@ def tw_mock():
             return text
 
         def get_write_msg(self, idx):
-            flag, msg = self.lines[idx]
-            assert flag == TWMock.WRITE
+            assert self.lines[idx][0] == TWMock.WRITE
+            msg = self.lines[idx][1]
             return msg
 
         fullwidth = 80
